+++ lib/generators/rails_settings/migration/templates/migration.rb
@@ -2,7 +2,7 @@ class RailsSettingsMigration < ActiveRecord::Migration
   def self.up
     create_table :settings do |t|
       t.string     :var,    :null => false
-      t.text       :value,  :null => false
+      t.text       :value
       t.references :target, :null => false, :polymorphic => true
       t.timestamps
     end